Fix main.cpp and makefile and add test for main function
Fix the main.cpp, stripping the \r or \n from input file.
Add flag into main.cpp and makefile so that when compiling there won't be main function conflict with test main function.
Test for main is using dup2 to simulate user input and the output of the function is dumped into null stream